+++ lib/main.dart
@@ -1,10 +1,9 @@
 import 'package:flutter/material.dart';
 import 'package:get/get_navigation/src/root/get_material_app.dart';
 import 'package:get_storage/get_storage.dart';
-import 'package:scheduler/note_alarm.dart';
 import 'package:scheduler/services/theme_services.dart';
 import 'package:scheduler/ui/home_page.dart';
-import 'package:scheduler/ui/notification_handler.dart';
+import 'package:scheduler/services/notification_services.dart';
 import 'package:scheduler/ui/theme.dart';
 import 'package:scheduler/ui/widgets/dynamicline.dart';
 import 'db/db_helper.dart';
