+++ lib/ui/widgets/task_tile.dart
@@ -1,7 +1,6 @@
 import 'package:flutter/material.dart';
 import 'package:google_fonts/google_fonts.dart';
 import 'package:scheduler/ui/add_task_bar.dart';
-
 import '../../models/task.dart';
 import '../theme.dart';
