+++ pubspec.yaml
@@ -48,6 +48,7 @@ dependencies:
   flutter_staggered_animations:
   flutter_colorpicker: ^1.0.3
   datetime_picker_formfield_new: ^2.0.0
+  number_to_words: ^1.0.0
 dev_dependencies:
   flutter_test:
     sdk: flutter
